Add v8 libbase dependency.
V8 has refactored some of its targets and we now need to link v8_libbase
as well as v8_base. Currently libbase has an architecture-dependent
name; until that's fixed upstream, we need to depend on the
appropriately named library, and we must enable grouping of static
libraries as otherwise the link order is incorrect and will fail.

Also, add external/chromium_org/src/v8 to the include path to handle the
chromium-dev branch source layout which is slightly different to master.

Change-Id: I26f71f4eaa7a2cf36a10e651212ce3214e129e7a